% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{compare_ranks}
\alias{compare_ranks}
\title{Transform ranks into comparisons}
\usage{
compare_ranks(data, nitems_in_block)
}
\arguments{
\item{data}{An object of class \code{data.frame} containing
all ranks for every block}

\item{nitems_in_block}{An integer indicating number of items inside one block.
This number must be more than 1.}
}
\description{
Transform dataframe of ranks into dataframe of item comparisons.
The resulted dataframe can be used in function \code{\link{make_TIRT_data}}
}
\examples{
#define a ranking data.frame with two rows and two blocks with three items each
ranks = data.frame(i1 = c(1,2), i2 = c(2,3), i3 = c(3,1),
                   i4 = c(3,1), i5 = c(1,3), i6 = c(2,2))

#generate comparisons data.frame
compare_ranks(data = ranks, nitems_in_block = 3)
}
\seealso{
\code{\link{make_TIRT_data}}
}
