% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Set up Correlation Matrices}
\usage{
cor_matrix(cors, dim, dimnames = NULL)
}
\arguments{
\item{cors}{vector of unique correlations}

\item{dim}{Dimension of the correlation matrix}

\item{dimnames}{Optional dimnames of the correlation matrix}
}
\value{
A correlation \code{matrix} of dimension \code{dim}.
}
\description{
Set up Correlation Matrices
}
\examples{
cor_matrix(c(0.2, 0.3, 0.5), dim = 3)

}
