% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.R
\name{fit_TIRT_lavaan}
\alias{fit_TIRT_lavaan}
\title{Fit Thurstonian IRT models in lavaan}
\usage{
fit_TIRT_lavaan(data, estimator = "ULSMV", ...)
}
\arguments{
\item{data}{An object of class \code{'TIRTdata'}. see
\code{\link{make_TIRT_data}} for documentation on how to create one.}

\item{estimator}{Name of the estimator that should be used.
See \code{\link[lavaan:lavOptions]{lavOptions}}.}

\item{...}{Further arguments passed to
\code{\link[lavaan:lavaan]{lavaan}}.}
}
\value{
A \code{'TIRTfit'} object.
}
\description{
Fit Thurstonian IRT models in lavaan
}
