% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.R
\name{fit_TIRT_mplus}
\alias{fit_TIRT_mplus}
\title{Fit Thurstonian IRT models in Mplus}
\usage{
fit_TIRT_mplus(data, ...)
}
\arguments{
\item{data}{An object of class \code{'TIRTdata'}. see
\code{\link{make_TIRT_data}} for documentation on how to create one.}

\item{...}{Further arguments passed to
\code{\link[MplusAutomation:mplusModeler]{mplusModeler}}.}
}
\value{
A \code{'TIRTfit'} object.
}
\description{
Fit Thurstonian IRT models in Mplus
}
