% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_TIRT_stan}
\alias{fit_TIRT_stan}
\title{Fit Thurstonian IRT models in Stan}
\usage{
fit_TIRT_stan(data, init = 0, ...)
}
\arguments{
\item{data}{An object of class \code{'TIRTdata'}. see
\code{\link{make_TIRT_data}} for documentation on how to create one.}

\item{init}{Initial values of the parameters.
Defaults to \code{0} as it proved to be most stable.}

\item{...}{Further arguments passed to
\code{\link[rstan:sampling]{rstan::sampling}}.}
}
\value{
A \code{'TIRTfit'} object.
}
\description{
Fit Thurstonian IRT models in Stan
}
