% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.R
\name{make_mplus_code}
\alias{make_mplus_code}
\title{Generate Mplus code for Thurstonian IRT models}
\usage{
make_mplus_code(data, blocks = NULL, iter = 1000,
  eta_file = "eta.csv")
}
\arguments{
\item{data}{An object of class \code{data.frame}
containing data of all variables used in the model.}

\item{blocks}{Object of class \code{TIRTblocks}
generated by \code{\link{set_block}} indicating
which items belong to which block, trait and more.
Ignored if data already contains information on
the blocks.}

\item{iter}{Maximum number of iterations of the
model fitting algorithm.}

\item{eta_file}{optional file name in which predicted
trait scores should be stored.}
}
\value{
A list of Mplus code snippets to be
interpreted by the \pkg{MplusAutomation} package.
}
\description{
Generate Mplus code for Thurstonian IRT models
}
\examples{
sdata <- sim_TIRT_data(
  npersons = 100,
  ntraits = 3,
  nblocks_per_trait = 4,
  gamma = 0,
  lambda = c(runif(6, 0.5, 1), runif(6, -1, -0.5)),
  Phi = diag(3)
)
lapply(make_mplus_code(sdata), cat)

}
