% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_thurstonian_data}
\alias{sim_thurstonian_data}
\title{Simulate Thurstonian IRT data}
\usage{
sim_thurstonian_data(npersons, ntraits, gamma, lambda, psi = NULL,
  Phi = NULL, eta = NULL, nblocks_per_trait = 5, nitems_per_block = 3,
  comb_blocks = c("fixed", "random"))
}
\arguments{
\item{npersons}{Number of persons}

\item{ntraits}{Number of traits}

\item{gamma}{intercept parameters}

\item{lambda}{item factor loadings}

\item{psi}{optional item uniquenesses}

\item{Phi}{optional trait correlation matrix}

\item{eta}{optional person factor scores}

\item{nblocks_per_trait}{Number of blocks per trait}

\item{nitems_per_block}{Number of items per block}

\item{comb_blocks}{Indicates how to combine traits
to blocks. \code{"fixed"} implies a simple non-random
design that may combine certain tratis which each
other disproportionally often. We thus recommend
to use a \code{"random"} block design that combines
all traits with all other traits equally often
on average.}
}
\value{
A \code{data.frame} of the same structure
as returned by \code{\link{make_TIRT_data}}.
}
\description{
Simulate Thurstonian IRT data
}
\examples{
sdata <- sim_thurstonian_data(
  npersons = 100,
  ntraits = 3,
  nblocks_per_trait = 4,
  gamma = 0,
  lambda = c(runif(6, 0.5, 1), runif(6, -1, -0.5)),
  Phi = diag(3)
)
head(sdata)

}
