\name{thurstonianIRT-package}
\alias{thurstonianIRT-package}
\alias{thurstonianIRT}
\docType{package}
\title{
\packageTitle{thurstonianIRT}
}
\description{
\packageDescription{thurstonianIRT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{thurstonianIRT}
\packageIndices{thurstonianIRT}
An overview of how to use the package, including the most important functions 
}
\author{
\packageAuthor{thurstonianIRT}

Maintainer: \packageMaintainer{thurstonianIRT}
}
\references{
Literature or other references for background information 
}
 Optionally other standard keywords, one per line, from file KEYWORDS in the R 
 documentation directory 
\keyword{ package }
\seealso{
 Optional links to other man pages, e.g. 
 \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
 simple examples of the most important functions 
}
