% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.R
\name{fit_TIRT_lavaan}
\alias{fit_TIRT_lavaan}
\title{Fit Thurstonian IRT models in lavaan}
\usage{
fit_TIRT_lavaan(data, blocks = NULL, estimator = "ULSMV", ...)
}
\arguments{
\item{data}{An object of class \code{data.frame} 
containing data of all variables used in the model.}

\item{blocks}{Object of class \code{TIRTblocks}
generated by \code{\link{set_block}} indicating
which items belong to which block, trait and more.
Ignored if data already contains information on 
the blocks.}

\item{estimator}{Name of the estimator that should be used. 
See \code{\link[lavaan:lavOptions]{lavOptions}}.}

\item{...}{Further arguments passed to 
\code{\link[lavaan:lavaan]{lavaan}}.}
}
\value{
A \code{TIRTfit} object.
}
\description{
Fit Thurstonian IRT models in lavaan
}
