% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.R
\name{fit_TIRT_mplus}
\alias{fit_TIRT_mplus}
\title{Fit Thurstonian IRT models in Mplus}
\usage{
fit_TIRT_mplus(data, blocks = NULL, ...)
}
\arguments{
\item{data}{An object of class \code{data.frame}
containing data of all variables used in the model.}

\item{blocks}{Object of class \code{TIRTblocks}
generated by \code{\link{set_block}} indicating
which items belong to which block, trait and more.
Ignored if data already contains information on
the blocks.}

\item{...}{Further arguments passed to
\code{\link[MplusAutomation:mplusModeler]{mplusModeler}}.}
}
\value{
A \code{TIRTfit} object.
}
\description{
Fit Thurstonian IRT models in Mplus
}
