% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_TIRT_stan}
\alias{fit_TIRT_stan}
\title{Fit Thurstonian IRT models in Stan}
\usage{
fit_TIRT_stan(data, blocks = NULL, init = 0, ...)
}
\arguments{
\item{data}{An object of class \code{data.frame} 
containing data of all variables used in the model.}

\item{blocks}{Object of class \code{TIRTblocks}
generated by \code{\link{set_block}} indicating
which items belong to which block, trait and more.
Ignored if data already contains information on 
the blocks.}

\item{init}{Initial values of the parameters.
Defaults to \code{0} as it proved to be most stable.}

\item{...}{Further arguments passed to 
\code{\link[rstan:sampling]{rstan::sampling}}.}
}
\value{
A \code{\link[rstan:stanfit-class]{stanfit}} object.
}
\description{
Fit Thurstonian IRT models in Stan
}
