% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_TIRT_stan}
\alias{fit_TIRT_stan}
\title{Fit Thurstonian IRT models in Stan}
\usage{
fit_TIRT_stan(data, init = 0, ...)
}
\arguments{
\item{data}{An object of class \code{'TIRTdata'}. see
\code{\link{make_TIRT_data}} for documentation on how to create one.}

\item{init}{Initial values of the parameters.
Defaults to \code{0} as it proved to be most stable.}

\item{...}{Further arguments passed to
\code{\link[rstan:sampling]{rstan::sampling}}.}
}
\value{
A \code{'TIRTfit'} object.
}
\description{
Fit Thurstonian IRT models in Stan
}
\examples{
# load the data
data("triplets")

# define the blocks of items
blocks <-
  set_block(c("i1", "i2", "i3"), traits = c("t1", "t2", "t3"),
          signs = c(1, 1, 1)) +
  set_block(c("i4", "i5", "i6"), traits = c("t1", "t2", "t3"),
            signs = c(-1, 1, 1)) +
  set_block(c("i7", "i8", "i9"), traits = c("t1", "t2", "t3"),
            signs = c(1, 1, -1)) +
  set_block(c("i10", "i11", "i12"), traits = c("t1", "t2", "t3"),
            signs = c(1, -1, 1))

# generate the data to be understood by 'thurstonianIRT'
triplets_long <- make_TIRT_data(
  data = triplets, blocks = blocks, direction = "larger",
  format = "pairwise", family = "bernoulli", range = c(0, 1)
)

\donttest{
# fit the data using Stan
fit <- fit_TIRT_stan(triplets_long, chains = 1)
print(fit)
predict(fit)
}

}
