% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{make_sem_data}
\alias{make_sem_data}
\title{Prepare data for Thurstonian IRT models fitted with
lavaan or Mplus}
\usage{
make_sem_data(data)
}
\arguments{
\item{data}{An object of class \code{'TIRTdata'}. see
\code{\link{make_TIRT_data}} for documentation on how to create one.}
}
\value{
A list of data ready to be passed to \pkg{lavaan}
or \pkg{Mplus}.
}
\description{
Prepare data for Thurstonian IRT models fitted with
lavaan or Mplus
}
