% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Prepare data for Thurstonian IRT models fitted with Stan}
\usage{
make_stan_data(data)
}
\arguments{
\item{data}{An object of class \code{data.frame} containing data of all
variables used in the model.}
}
\value{
A list of data ready to be passed to \pkg{Stan}.
}
\description{
Prepare data for Thurstonian IRT models fitted with Stan
}
