% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{make_TIRT_data}
\alias{make_TIRT_data}
\title{Prepare data for Thurstonian IRT models}
\usage{
make_TIRT_data(data, blocks, direction = c("smaller", "larger"),
  format = c("ranks", "pairwise"), family = "bernoulli",
  partial = FALSE, range = NULL)
}
\arguments{
\item{data}{An object of class \code{data.frame} containing data of all
variables used in the model.}

\item{blocks}{Object of class \code{TIRTblocks} generated by
\code{\link{set_block}} indicating which items belong to which block, trait
and more. Ignored if data already contains information on the blocks.}

\item{direction}{Indicates if \code{"smaller"} (the default) or
\code{"larger"} input values are considered as indicating the favored
answer.}

\item{format}{Format of the item responses. Either \code{"ranks"} for
responses in the form of ranks of \code{"pairwise"} for pairwise
responses.}

\item{family}{Name of assumed the response distribution. Either
\code{"bernoulli"}, \code{"cumulative"}, or \code{"gaussian"}.}

\item{partial}{A flag to indicate whether partial comparisons are allowed
for responses stored in the \code{"ranks"} format.}
}
\value{
A \code{data.frame} in a specific format and with attributes ready
  for use with other functions of the \pkg{ThurstonianIRT} package.
}
\description{
Prepare data for Thurstonian IRT models
}
