% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{make_TIRT_data}
\alias{make_TIRT_data}
\title{Prepare data for Thurstonian IRT models}
\usage{
make_TIRT_data(data, blocks)
}
\arguments{
\item{data}{An object of class \code{data.frame} 
containing data of all variables used in the model.}

\item{blocks}{Object of class \code{TIRTblocks}
generated by \code{\link{set_block}} indicating
which items belong to which block, trait and more.
Ignored if data already contains information on 
the blocks.}
}
\value{
A \code{data.frame} in a specific format
and with attributes ready for use with other
functions of the \pkg{ThurstonianIRT} package.
}
\description{
Prepare data for Thurstonian IRT models
}
