% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thurstonianIRT-package.R
\docType{package}
\name{thurstonianIRT-package}
\alias{thurstonianIRT-package}
\alias{thurstonianIRT}
\title{The 'thurstonianIRT' package.}
\description{
This package fits Thurstonian IRT models using Stan,
lavaan, and Mplus.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
